flask-restplus framework structure
