+++ apps/api/restplus.py
@@ -1,0 +1,3 @@
+from flask_restplus import Api
+
+api = Api()
